requirements.txt
streamlit
pandas
scikit-learn
xgboost
rdkit-pypi
joblib